Added webpack bundle and husky git hooks
