+++ test/fixtures/asyncSum.js
@@ -1,16 +1,14 @@
-module.exports = (a, b) => {
-    return new Promise((resolve, reject) => {
-        const saneA = Number.parseInt(a, 10);
-        const saneB = Number.parseInt(b, 10);
+module.exports = (a, b) => new Promise((resolve, reject) => {
+  const saneA = Number.parseInt(a, 10);
+  const saneB = Number.parseInt(b, 10);
 
-        if(isNaN(saneA)) {
-            reject('First argument is not number')
-        }
+  if (Number.isNaN(saneA)) {
+    reject(new TypeError('First argument is not number'));
+  }
 
-        if(isNaN(saneB)) {
-            reject('Second argument is not number')
-        }
+  if (Number.isNaN(saneB)) {
+    reject(new TypeError('Second argument is not number'));
+  }
 
-        resolve(saneA + saneB);
-    })
-}
+  resolve(saneA + saneB);
+});
