+++ test/index.js
@@ -1,0 +1,5 @@
+require('./async-node.specs');
+require('./async-sum.specs');
+require('./async-timeout.specs');
+require('./suite-setup-teardown.specs');
+require('./sum.specs');
